check for running container

+++ cicd/scripts/stop-container.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
 ssh -o "StrictHostKeyChecking no" -i key/dillo-key.pem $USER@$HOST <<- ENDSSH
-    docker kill dillo_bot
+    if [[ $(docker ispect dillo_bot | grep "Running" | grep true) ]]; then
+        docker kill dillo_bot
+    fi
 
     exit
 ENDSSH
